#25 [RENAME] 컨벤션에 맞게 폴더 구조 및 파일명 수정
