+++ src/index.js
@@ -7,8 +7,8 @@ import './index.css';
 import AboutPage from './pages/AboutPage/AboutPage';
 import AlertPage from './pages/AlertPage/AlertPage';
 import AuthPage from './pages/AuthPage/AuthPage';
-import BoardListPage from './pages/BoardListPage/BoardListPage';
-import BoardPage from './pages/BoardPage/BoardPage';
+import { BoardPage, BoardListPage } from './pages/BoardPage';
+import { PostPage, PostSearchPage, PostWritePage } from './pages/PostPage';
 import ErrorPage from './pages/ErrorPage/ErrorPage';
 import ExamReviewPage from './pages/ExamReviewPage/ExamReviewPage';
 import HelpPage from './pages/HelpPage/HelpPage';
@@ -32,6 +32,9 @@ const router = createBrowserRouter([
       { path: '/board/large-snow', element: <BoardListPage /> },
       { path: '/board/permanent-snow', element: <BoardListPage /> },
       { path: '/board/besookt', element: <BoardListPage /> },
+      { path: '/post', element: <PostPage /> },
+      { path: '/post-search', element: <PostSearchPage /> },
+      { path: '/post-write', element: <PostWritePage /> },
       { path: 'exam-review', element: <ExamReviewPage /> },
       { path: 'alert', element: <AlertPage /> },
       { path: 'my-page', element: <MyPage /> },
