+++ src/pages/BoardPage/index.js
@@ -1,0 +1,2 @@
+export { default as BoardPage } from './BoardPage/BoardPage';
+export { default as BoardListPage } from './BoardListPage/BoardListPage';
