--- src/pages/PostListPage/PostListPage.module.css
@@ -1,125 +1,0 @@
-body {
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-}
-::-webkit-scrollbar {
-  display: none;
-}
-
-.container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 8px;
-}
-
-.refresh_icon {
-  display: flex;
-  justify-content: center;
-  width: 40px;
-  height: 50px;
-  top: 10px;
-  position: fixed;
-}
-
-.pencil_icon {
-  display: flex;
-  justify-content: end;
-  width: 100%;
-  max-width: 450px;
-  background-size: cover;
-  position: fixed;
-  bottom: 160px;
-  margin-right: 40px;
-
-  :first-child {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-.top {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 8px;
-  /* 양옆 여백 */
-  padding: 0 17px;
-  border-bottom: 1px solid gray;
-}
-
-.header {
-  height: 42px;
-  display: flex;
-  align-items: center;
-  width: 100%;
-
-  h1 {
-    font-size: 12px;
-    font-weight: 700;
-    padding-left: 6px;
-  }
-}
-
-.side_menu_btn {
-  width: 20px;
-  height: 15px;
-  display: flex;
-}
-
-.notification_bar {
-  width: 100%;
-  background-color: #fee9eb;
-  height: 20px;
-  border-radius: 7px;
-  display: flex;
-  align-items: center;
-}
-
-.notification_icon {
-  width: 20px;
-  height: 15px;
-  display: flex;
-}
-
-.keyword_box {
-  display: flex;
-  width: 100%;
-  padding: 10px 0 10px 0;
-  justify-content: baseline;
-}
-
-.keyword {
-  height: 12px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 7px 5px;
-  border-radius: 8px;
-  border: 1px solid #5f86bf;
-  background: #ddebf6;
-  margin-right: 5px;
-  color: #5f86bf;
-  font-weight: 600;
-}
-
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 8px;
-  /* 양옆 여백 */
-  padding: 0 17px;
-}
-
-.sponser {
-  display: flex;
-  width: 100%;
-  max-width: 450px;
-  background-size: cover;
-  position: fixed;
-  bottom: 46px;
-}
