+++ src/pages/PostPage/index.js
@@ -1,0 +1,3 @@
+export { default as PostPage } from './PostPage/PostPage';
+export { default as PostSearchPage } from './PostSearchPage/PostSearchPage';
+export { default as PostWritePage } from './PostWritePage/PostWritePage';
